etl-final/detector-service/requirements.txt
Django
djangorestframework
kafka-python
requests
pymysql
psycopg2-binary
python-dotenv
